[SDL] Revert to unconditionally drawing the entire screen
286c4814f97eb0d06195ef492d8906a417195e52 introduced a regression on
some GNU/Linux systems, whereas the screen would flicker with the
contents of previous screens or uninitialized video memory. The
working theory for the root cause of this glitch is that SDL does not
always retain the previous screen contents across SDL_RenderPresent
calls, thus the above-mentioned patch occasionally chose to not redraw
the screen when a redraw was actually necessary.

For now, revert this logic and redraw the screen unconditionally
again, until a better way to improve CPU usage is found.