Replace unportable wmemchr() call with portable code
wchar_t is not 16 bits in size on all platforms.

I only used originally used wmemchr() because I assumed that in x86 it would compile to "REPNE SCASW" (the same as in the original DOS binary) when the compiler was told to optimize for size, but existing compilers apparently do not do that. So there's no good reason not to just use an easy-to-understand loop instead of an awkward library call.

Fixes issue #15.