[TA] Switch from playback to rerecording only when Ctrl+T is pressed
This now allows replays to be rewound without modifying them. There is currently no visible feedback upon switching modes, though.

TODO: Extend the replay format to include a rerecord count (this would be the number of times active play was resumed after stepping backwards by any number of frames).
And when that is done, fields for whether to emulate/fix bugs should be added too, with room for future expansion.
Ideally, there would be disambiguation for whether a file is in the new or old format.