[SDL] Optimize away rendering empty glyphs and blackground
The alpha-only rendering is slower than the previous (pre-afc49043bf4730967b286c20dc49e54f3e9c8be7) method, causing it not to be able to render at full framerate with CHEAT_OMNISCIENCE on my machine, but this optimization is enough to bring it back up to full framerate.

It's still unsatisfying for the rendering to not be as fast as it could be. A shader could probably do this efficiently, and I'll probably try that. If that doesn't pan out, I might switch back to the previous method, where each character+foreground+background combination has its own texture.